+++ test/performance/putTokenPdv.js
@@ -6,10 +6,10 @@ import {
 import { assert, statusOk, } from '../common/assertions.js'
 import { isEnvValid, DEV, UAT, PROD } from '../common/envs.js'
 import { getFCList } from '../common/utils.js'
-import {vu} from 'k6/execution'
+import {exec, vu, scenario} from 'k6/execution'
 import { SharedArray } from 'k6/data'
 import { jUnit, textSummary } from 'https://jslib.k6.io/k6-summary/0.0.2/index.js';
-import { setStages, setScenarios } from '../common/stageUtils.js';
+import { setStages, setScenarios, buildScenarios, coalesce } from '../common/stageUtils.js';
 import defaultHandleSummaryBuilder from '../common/handleSummaryBuilder.js'
 
 const REGISTERED_ENVS = [DEV, UAT, PROD]
@@ -63,26 +63,6 @@ export let options = typeScenario
 
 if (isEnvValid(__ENV.TARGET_ENV)) {
     baseUrl = services[`${__ENV.TARGET_ENV}_pdv`].baseUrl
-}
-
-function buildScenarios(options) {
-    let counter = 0
-    const scenarioBaseIndexes = {}
-
-    Object.keys(options.scenarios)
-        .filter(scenarioName => scenarioName.startsWith('scenario_'))
-        .sort()
-        .forEach(scenarioName => {
-            const singleScenario = options.scenarios[scenarioName]
-            let scenarioBaseIndex = counter
-            counter += singleScenario.vus
-            scenarioBaseIndexes[scenarioName] = scenarioBaseIndex
-        })
-    return scenarioBaseIndexes
-}
-
-function coalesce(o1, o2){
-    return o1 ? o1 : o2
 }
 
 export default () => {
